+++ src/Service/Yaml/FilesFinder.php
@@ -5,16 +5,19 @@ declare(strict_types=1);
 namespace Aeliot\Bundle\TransMaintain\Service\Yaml;
 
 use Aeliot\Bundle\TransMaintain\Service\DirectoryProvider;
+use Aeliot\Bundle\TransMaintain\Service\FileMapBuilder;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
 
 final class FilesFinder
 {
     private DirectoryProvider $directoryProvider;
+    private FileMapBuilder $fileMapBuilder;
 
-    public function __construct(DirectoryProvider $directoryProvider)
+    public function __construct(DirectoryProvider $directoryProvider, FileMapBuilder $fileMapBuilder)
     {
         $this->directoryProvider = $directoryProvider;
+        $this->fileMapBuilder = $fileMapBuilder;
     }
 
     /**
@@ -33,21 +36,7 @@ final class FilesFinder
      */
     public function getFilesMap(): array
     {
-        $map = [];
-        foreach ($this->getFiles() as $file) {
-            $parts = explode('.', $file->getFilenameWithoutExtension());
-            $locale = array_pop($parts);
-            $domain = implode('.', $parts);
-            if (!isset($map[$domain])) {
-                $map[$domain] = [];
-            }
-            if (!isset($map[$domain][$locale])) {
-                $map[$domain][$locale] = [];
-            }
-            $map[$domain][$locale][] = $file->getRealPath();
-        }
-
-        return $map;
+        return $this->fileMapBuilder->buildFilesMap($this->getFiles());
     }
 
     /**
@@ -77,7 +66,7 @@ final class FilesFinder
     }
 
     /**
-     * @return Finder|iterable<SplFileInfo>
+     * @return Finder<SplFileInfo>
      */
     private function getFiles(): Finder
     {